file uploading and downloading

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require("../models/product");
 const {validationResult} = require('express-validator');
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     // if(!req.session.isLoggedIn){
@@ -24,11 +25,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+   if(!image) {
+       return res.status(422).render('admin/edit-product', {
+           pageTitle: 'Add Product',
+           path: '/admin/add-product',
+           editing: false,
+           product: null,
+           errorMessage: 'Attached file is not an image!',
+           oldInput: {
+               title: title,
+               price: price,
+               description: description,
+           },
+           validationErrors: [],
+       });
+   }
+    console.log(image);
+   const imageUrl = image.path;
     const errors = validationResult(req);
-
     if(!errors.isEmpty()) {
         console.log(errors.array());
         return res.status(422).render('admin/edit-product', {
@@ -39,7 +56,6 @@ exports.postAddProduct = (req, res, next) => {
             errorMessage: errors.array()[0].msg,
             oldInput: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description,
             },
@@ -88,7 +104,6 @@ exports.getEditProduct = (req, res, next) => {
             errorMessage: '',
             oldInput: {
                 title: '',
-                imageUrl: '',
                 price: 0,
                 description: '',
             },
@@ -106,7 +121,7 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDescription = req.body.description;
     const errors = validationResult(req);
 
@@ -118,7 +133,6 @@ exports.postEditProduct = (req, res, next) => {
             editing: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: prodId,
@@ -136,7 +150,10 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDescription;
-            product.imageUrl = updatedImageUrl;
+            if(image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             return product.save()
                 .then(result => {
                     console.log('Product updated');
@@ -152,12 +169,21 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
-        .then(() => {
+    Product.findById(prodId).then(product => {
+        if(!product) {
+            return next(new Error('Product not found!'))
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return  Product.deleteOne({_id: prodId, userId: req.user._id})
+    }).then(() => {
             console.log('Product deleted');
             res.redirect('/admin/products');
         })
-        .catch(err => console.log(err));
+        .catch(err => {
+            const error =  new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        });
 };
 
 exports.getProducts = (req, res, next) => {
